Drought-related index functions

The functions are used for quick computing of some common drought indexes built from wrappers of the underlying modules. For more customised needs, users may build their own indexes from the modules.

Usage

trans_thornthwaite(var, lat, na.rm = FALSE, verbose = TRUE)

idx_spi(data, .prcp, .dist = dist_gamma(), .scale = 12)

idx_spei(
  data,
  .tavg,
  .lat,
  .prcp,
  .pet_method = trans_thornthwaite(),
  .scale = 12,
  .dist = dist_glo()
)

idx_rdi(
  data,
  .tavg,
  .lat,
  .prcp,
  .pet_method = trans_thornthwaite(),
  .scale = 12
)

idx_edi(data, .tavg, .lat, .prcp, .scale = 12)

Arguments

var

the variable to be transformed, see [tidyindex::variable_trans()] and [SPEI::thornthwaite()]

lat, na.rm, verbose

see [SPEI::thornthwaite]

data

an id_tbl object

.dist

the distribution used for distribution fit, see [tidyindex::distribution_fit()]

.scale

the temporal aggregation scale, see [tidyindex::temporal_aggregation()]

.tavg, .lat, .prcp

variables to be used in the index calculation, see Details

.pet_method

the method used for calculating potential evapotranspitation, currently only trans_thornthwaite()

Details

Below explains the steps wrapped in each index and the intermediate variables created.

The idx_spi() function performs

a temporal aggregation on the input precipitation series, .prcp, as .agg,

adistribution fit step on the aggregated precipitation , .agg, as .fit, and

a normalising step on the fitted values, .fit, as .index

The idx_spei() function performs

a variable transformation step on the inut average temperature, .tavg, to obtain the potential evapotranspiration, .pet,

a dimension reduction step to calculate difference series, .diff, between the input precipitation series, .prcp, and .pet,

a temporal aggregation step on the difference series, .diff, as .agg,

a distribution fit step on the aggregated series, .agg, as .fit, and

a normalising step on the fitted value, .fit, to obtain .index.

The idx_rdi() function performs

a variable transformation step on the input average temperature, .tavg, to obtain potential evapotranspiration .pet,

a dimension reduction step to calculate the ratio of input precipitation, .prcp, to .pet as .ratio,

a temporal aggregation step on the ratio series, .ratio, as .agg

a variable transformation step to take the log10 of the aggregated series, .agg, as .y, and

a rescaling step to rescale .y by zscore to obtain .index.

The idx_edi() function performs

a dimension reduction step to aggregate the input precipitation series, prcp, as .mult,

a temporal aggregation step on the aggregated precipitation series (.mult) as .ep, and

a rescaling step to escale .ep by zscore to obtain .index.

Examples

library(dplyr)
library(lmomco)
dt <- tenterfield |>
  mutate(month = lubridate::month(ym)) |>
  init(id = id, time = ym, group = month)

dt |> idx_spi()
#> Index pipeline: 
#> 
#> Steps: 
#> temporal: `rolling_window()` -> .agg
#> distribution_fit: `distfit_gamma()` -> .fit
#> normalise: `norm_quantile()` -> .index
#> 
#> Data: 
#> # A tibble: 369 × 14
#>    id           ym  prcp  tmax  tmin  tavg  long   lat name  month  .agg    .fit
#>    <chr>     <mth> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <chr> <dbl> <dbl>   <dbl>
#>  1 ASN00… 1990 Jan   882  27.0 15.2  21.1   152. -29.0 tent…     1   882 7.25e-8
#>  2 ASN00… 1990 Feb  1260  26.1 16.0  21.0   152. -29.0 tent…     2  2142 7.38e-1
#>  3 ASN00… 1990 Mar   254  23.8 13.4  18.6   152. -29.0 tent…     3  2396 3.15e-1
#>  4 ASN00… 1990 Apr  1594  20.4 12.5  16.5   152. -29.0 tent…     4  3990 8.12e-1
#>  5 ASN00… 1990 May  1220  19.1  6.66 12.9   152. -29.0 tent…     5  5210 2.40e-1
#>  6 ASN00… 1990 Jun   394  14.6  3.19  8.88  152. -29.0 tent…     6  5604 4.04e-1
#>  7 ASN00… 1990 Jul   618  15.5  1.95  8.75  152. -29.0 tent…     7  6222 8.36e-1
#>  8 ASN00… 1990 Aug   334  14.3  2.49  8.41  152. -29.0 tent…     8  6556 8.25e-1
#>  9 ASN00… 1990 Sep   266  18.7  5.4  12.1   152. -29.0 tent…     9  6822 7.36e-1
#> 10 ASN00… 1990 Oct   362  23.3  7.6  15.4   152. -29.0 tent…    10  7184 7.58e-1
#> # ℹ 359 more rows
#> # ℹ 2 more variables: .fit_obj <list>, .index <dbl>
dt |> idx_spi(.scale = c(12, 24))
#> Index pipeline: 
#> 
#> Steps: 
#> temporal: `rolling_window()` -> .agg_12
#> temporal: `rolling_window()` -> .agg_24
#> distribution_fit: `distfit_gamma()` -> .fit_12
#> distribution_fit: `distfit_gamma()` -> .fit_24
#> normalise: `norm_quantile()` -> .index_12
#> normalise: `norm_quantile()` -> .index_24
#> 
#> Data: 
#> # A tibble: 369 × 18
#>    id                ym  prcp  tmax  tmin  tavg  long   lat name   month .agg_12
#>    <chr>          <mth> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <chr>  <dbl>   <dbl>
#>  1 ASN00056032 1990 Jan   882  27.0 15.2  21.1   152. -29.0 tente…     1     882
#>  2 ASN00056032 1990 Feb  1260  26.1 16.0  21.0   152. -29.0 tente…     2    2142
#>  3 ASN00056032 1990 Mar   254  23.8 13.4  18.6   152. -29.0 tente…     3    2396
#>  4 ASN00056032 1990 Apr  1594  20.4 12.5  16.5   152. -29.0 tente…     4    3990
#>  5 ASN00056032 1990 May  1220  19.1  6.66 12.9   152. -29.0 tente…     5    5210
#>  6 ASN00056032 1990 Jun   394  14.6  3.19  8.88  152. -29.0 tente…     6    5604
#>  7 ASN00056032 1990 Jul   618  15.5  1.95  8.75  152. -29.0 tente…     7    6222
#>  8 ASN00056032 1990 Aug   334  14.3  2.49  8.41  152. -29.0 tente…     8    6556
#>  9 ASN00056032 1990 Sep   266  18.7  5.4  12.1   152. -29.0 tente…     9    6822
#> 10 ASN00056032 1990 Oct   362  23.3  7.6  15.4   152. -29.0 tente…    10    7184
#> # ℹ 359 more rows
#> # ℹ 7 more variables: .agg_24 <dbl>, .fit_12 <dbl>, .fit_12_obj <list>,
#> #   .fit_24 <dbl>, .fit_24_obj <list>, .index_12 <dbl>, .index_24 <dbl>
dt |> idx_spei(.lat = lat, .tavg = tavg)
#> [1] "Checking for missing values (`NA`): all the data must be complete. Input type is vector. Assuming the data are monthly time series starting in January, all regular (non-leap) years."
#> Index pipeline: 
#> 
#> Steps: 
#> variable_transformation: `trans_thornthwaite()` -> .pet
#> dimension_reduction: `aggregate_manual()` -> .diff
#> temporal: `rolling_window()` -> .agg
#> distribution_fit: `distfit_glo()` -> .fit
#> normalise: `norm_quantile()` -> .index
#> 
#> Data: 
#> # A tibble: 369 × 16
#>    id             ym  prcp  tmax  tmin  tavg  long   lat name  month  .pet .diff
#>    <chr>       <mth> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <chr> <dbl> <dbl> <dbl>
#>  1 ASN0005… 1990 Jan   882  27.0 15.2  21.1   152. -29.0 tent…     1 113.   769.
#>  2 ASN0005… 1990 Feb  1260  26.1 16.0  21.0   152. -29.0 tent…     2  97.0 1163.
#>  3 ASN0005… 1990 Mar   254  23.8 13.4  18.6   152. -29.0 tent…     3  83.9  170.
#>  4 ASN0005… 1990 Apr  1594  20.4 12.5  16.5   152. -29.0 tent…     4  62.8 1531.
#>  5 ASN0005… 1990 May  1220  19.1  6.66 12.9   152. -29.0 tent…     5  42.1 1178.
#>  6 ASN0005… 1990 Jun   394  14.6  3.19  8.88  152. -29.0 tent…     6  22.5  372.
#>  7 ASN0005… 1990 Jul   618  15.5  1.95  8.75  152. -29.0 tent…     7  23.1  595.
#>  8 ASN0005… 1990 Aug   334  14.3  2.49  8.41  152. -29.0 tent…     8  23.1  311.
#>  9 ASN0005… 1990 Sep   266  18.7  5.4  12.1   152. -29.0 tent…     9  41.2  225.
#> 10 ASN0005… 1990 Oct   362  23.3  7.6  15.4   152. -29.0 tent…    10  66.1  296.
#> # ℹ 359 more rows
#> # ℹ 4 more variables: .agg <dbl>, .fit <dbl>, .fit_obj <list>, .index <dbl>
dt |> idx_rdi(.lat = lat, .tavg = tavg)
#> [1] "Checking for missing values (`NA`): all the data must be complete. Input type is vector. Assuming the data are monthly time series starting in January, all regular (non-leap) years."
#> Index pipeline: 
#> 
#> Steps: 
#> variable_transformation: `trans_thornthwaite()` -> .pet
#> dimension_reduction: `aggregate_manual()` -> .ratio
#> temporal: `rolling_window()` -> .agg
#> variable_transformation: `trans_log()` -> .y
#> rescaling: `rescale_zscore()` -> .index
#> 
#> Data: 
#> # A tibble: 369 × 15
#>    id            ym  prcp  tmax  tmin  tavg  long   lat name  month  .pet .ratio
#>    <chr>      <mth> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <chr> <dbl> <dbl>  <dbl>
#>  1 ASN000… 1990 Jan   882  27.0 15.2  21.1   152. -29.0 tent…     1 113.    7.80
#>  2 ASN000… 1990 Feb  1260  26.1 16.0  21.0   152. -29.0 tent…     2  97.0  13.0 
#>  3 ASN000… 1990 Mar   254  23.8 13.4  18.6   152. -29.0 tent…     3  83.9   3.03
#>  4 ASN000… 1990 Apr  1594  20.4 12.5  16.5   152. -29.0 tent…     4  62.8  25.4 
#>  5 ASN000… 1990 May  1220  19.1  6.66 12.9   152. -29.0 tent…     5  42.1  29.0 
#>  6 ASN000… 1990 Jun   394  14.6  3.19  8.88  152. -29.0 tent…     6  22.5  17.5 
#>  7 ASN000… 1990 Jul   618  15.5  1.95  8.75  152. -29.0 tent…     7  23.1  26.8 
#>  8 ASN000… 1990 Aug   334  14.3  2.49  8.41  152. -29.0 tent…     8  23.1  14.5 
#>  9 ASN000… 1990 Sep   266  18.7  5.4  12.1   152. -29.0 tent…     9  41.2   6.45
#> 10 ASN000… 1990 Oct   362  23.3  7.6  15.4   152. -29.0 tent…    10  66.1   5.48
#> # ℹ 359 more rows
#> # ℹ 3 more variables: .agg <dbl>, .y <dbl>, .index <dbl>
dt |> idx_edi(.lat = lat, .tavg = tavg)
#> Index pipeline: 
#> 
#> Steps: 
#> dimension_reduction: `aggregate_manual()` -> .mult
#> temporal: `rolling_window()` -> .ep
#> rescaling: `rescale_zscore()` -> .index
#> 
#> Data: 
#> # A tibble: 369 × 13
#>    id           ym  prcp  tmax  tmin  tavg  long   lat name  month  .mult    .ep
#>    <chr>     <mth> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <chr> <dbl>  <dbl>  <dbl>
#>  1 ASN00… 1990 Jan   882  27.0 15.2  21.1   152. -29.0 tent…     1  5724.  5724.
#>  2 ASN00… 1990 Feb  1260  26.1 16.0  21.0   152. -29.0 tent…     2  8173. 13897.
#>  3 ASN00… 1990 Mar   254  23.8 13.4  18.6   152. -29.0 tent…     3  1647. 15544.
#>  4 ASN00… 1990 Apr  1594  20.4 12.5  16.5   152. -29.0 tent…     4 10331. 25875.
#>  5 ASN00… 1990 May  1220  19.1  6.66 12.9   152. -29.0 tent…     5  7904. 33779.
#>  6 ASN00… 1990 Jun   394  14.6  3.19  8.88  152. -29.0 tent…     6  2551. 36330.
#>  7 ASN00… 1990 Jul   618  15.5  1.95  8.75  152. -29.0 tent…     7  4000. 40330.
#>  8 ASN00… 1990 Aug   334  14.3  2.49  8.41  152. -29.0 tent…     8  2161. 42491.
#>  9 ASN00… 1990 Sep   266  18.7  5.4  12.1   152. -29.0 tent…     9  1720. 44212.
#> 10 ASN00… 1990 Oct   362  23.3  7.6  15.4   152. -29.0 tent…    10  2340. 46552.
#> # ℹ 359 more rows
#> # ℹ 1 more variable: .index <dbl>