The scaling module

Specify either as an expression (like mutate) or using .method.

Usage

rescaling(data, ..., .method = NULL, .vars = NULL, .new_name = NULL)

rescale_zscore(var, na.rm = TRUE)

rescale_minmax(var, min = NULL, max = NULL, na.rm = TRUE, censor = TRUE)

rescale_center(var, na.rm = TRUE)

Arguments

data

an index table object

...

expression

.method

the method use for rescaling, currently rescale_minmax, rescale_zscore, rescale_center

.vars

variables to rescale

.new_name

the new name for the rescaled variable, optional

var

variable, argument in the rescale_* functions

na.rm

logical; whether to remove NAs, argument in the rescale_* functions

min, max

the minimum and maximum value used in rescale_minmax

censor

logical; whether to censor points outside min and max, if provided

Details

When using .method, specifies the varaibles it applies to (`.vars`), can be multiple, use tidyselect syntax additional parameters will be parsed from `...` `.new_names` needs to be of same length as `.vars`

Examples

dt <- hdi %>% init(id = country, indicators = life_exp:gni_pc)
#> Joining with `by = join_by(variables)`
# only print out the operation table
show_summary <- function(obj) obj %>% .[["op"]]

# single variable:
show_summary(dt %>% rescaling(life_exp = rescale_minmax(life_exp)))
#> # A tibble: 1 × 4
#>   module    step    var                      res     
#>   <chr>     <chr>   <chr>                    <chr>   
#> 1 rescaling formula rescale_minmax(life_exp) life_exp
show_summary(dt %>% rescaling(.method = rescale_minmax, .vars = life_exp, min = 20, max = 85))
#> # A tibble: 1 × 4
#>   module    step           var                                             res  
#>   <chr>     <chr>          <chr>                                           <chr>
#> 1 rescaling rescale_minmax rescale_minmax(var = life_exp, min = 20, max =… life…
show_summary(dt %>% rescaling(.method = rescale_minmax, .vars = life_exp, .new_name = "life_exp2"))
#> # A tibble: 1 × 4
#>   module    step           var                            res      
#>   <chr>     <chr>          <chr>                          <chr>    
#> 1 rescaling rescale_minmax rescale_minmax(var = life_exp) life_exp2

# use tidyselect to apply the .method to multiple variables
show_summary(dt %>% rescaling(.method = rescale_minmax, .vars = life_exp:exp_sch))
#> # A tibble: 2 × 4
#>   module    step           var                            res     
#>   <chr>     <chr>          <chr>                          <chr>   
#> 1 rescaling rescale_minmax rescale_minmax(var = life_exp) life_exp
#> 2 rescaling rescale_minmax rescale_minmax(var = exp_sch)  exp_sch 
# can also supply additional parameters
show_summary(dt %>% rescaling(.method = rescale_minmax,
                 .vars = life_exp:exp_sch,
                 min = c(20, 0), max = c(85, 18),
                 .new_name = c("life_exp2", "exp_sch")))
#> # A tibble: 2 × 4
#>   module    step           var                                             res  
#>   <chr>     <chr>          <chr>                                           <chr>
#> 1 rescaling rescale_minmax rescale_minmax(var = life_exp, min = 20, max =… life…
#> 2 rescaling rescale_minmax rescale_minmax(var = exp_sch, min = 0, max = 1… exp_…